requirements.txt
dj-database-url==0.3.0
Django==1.8.6
django-dotenv==1.3.0
psycopg2==2.6.1
djangorestframework==3.3.1
Pillow==3.0.0
https://github.com/jezdez/django-configurations/archive/5ece1070448d3e026b81422691c60207f5699c0c.zip#egg=django-configurations==0.8.0.1
